Dead code cleanup
