+++ packages/shared.test/src/index.ts
@@ -18,7 +18,6 @@ export { authHandlerMiddleware } from "./handlers/authHandlerMiddleware.js";
 export { MockOntologiesV2 } from "./handlers/MockOntologiesV2.js";
 export { handleOpenApiCall } from "./handlers/util/handleOpenApiCall.js";
 export { loadAll } from "./handlers/util/loadAll.js";
-export * from "./mock-ontology/index.js";
 export { startNodeApiServer } from "./setupServers.js";
 export type { TestSetup } from "./setupServers.js";
 export * as stubData from "./stubs/index.js";
