--- packages/shared.test/src/mock-ontology/ObjectTypeWithAllPropertyTypes.ts
@@ -1,292 +1,0 @@
-/*
- * Copyright 2023 Palantir Technologies, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { ObjectMetadata, VersionBound } from "@osdk/api";
-
-export interface ObjectTypeWithAllPropertyTypesDef
-  extends ObjectMetadata, VersionBound<"0.15.0">
-{
-  type: "object";
-  apiName: "ObjectTypeWithAllPropertyTypes";
-  primaryKeyApiName: "id";
-  primaryKeyType: "integer";
-  links: {};
-  properties: {
-    id: {
-      multiplicity: false;
-      type: "integer";
-    };
-    string: {
-      multiplicity: false;
-      type: "string";
-    };
-    boolean: {
-      multiplicity: false;
-      type: "boolean";
-    };
-    date: {
-      multiplicity: false;
-      type: "datetime";
-    };
-    dateTime: {
-      multiplicity: false;
-      type: "timestamp";
-    };
-    decimal: {
-      multiplicity: false;
-      type: "decimal";
-    };
-    integer: {
-      multiplicity: false;
-      type: "integer";
-    };
-    long: {
-      multiplicity: false;
-      type: "long";
-    };
-    short: {
-      multiplicity: false;
-      type: "short";
-    };
-    float: {
-      multiplicity: false;
-      type: "float";
-    };
-    double: {
-      multiplicity: false;
-      type: "double";
-    };
-    byte: {
-      multiplicity: false;
-      type: "byte";
-    };
-    attachment: {
-      multiplicity: false;
-      type: "attachment";
-    };
-    geoPoint: {
-      multiplicity: false;
-      type: "geopoint";
-    };
-    geoShape: {
-      multiplicity: false;
-      type: "geoshape";
-    };
-    stringArray: {
-      multiplicity: true;
-      type: "string";
-    };
-    booleanArray: {
-      multiplicity: true;
-      type: "boolean";
-    };
-    dateArray: {
-      multiplicity: true;
-      type: "datetime";
-    };
-    dateTimeArray: {
-      multiplicity: true;
-      type: "timestamp";
-    };
-    decimalArray: {
-      multiplicity: true;
-      type: "decimal";
-    };
-    integerArray: {
-      multiplicity: true;
-      type: "integer";
-    };
-    longArray: {
-      multiplicity: true;
-      type: "long";
-    };
-    shortArray: {
-      multiplicity: true;
-      type: "short";
-    };
-    floatArray: {
-      multiplicity: true;
-      type: "float";
-    };
-    doubleArray: {
-      multiplicity: true;
-      type: "double";
-    };
-    byteArray: {
-      multiplicity: true;
-      type: "byte";
-    };
-    attachmentArray: {
-      multiplicity: true;
-      type: "attachment";
-    };
-    geoPointArray: {
-      multiplicity: true;
-      type: "geopoint";
-    };
-    geoShapeArray: {
-      multiplicity: true;
-      type: "geoshape";
-    };
-    mediaReference: {
-      multiplicity: false;
-      type: "mediaReference";
-    };
-  };
-}
-
-export const ObjectTypeWithAllPropertyTypes: ObjectTypeWithAllPropertyTypesDef =
-  {
-    type: "object",
-    apiName: "ObjectTypeWithAllPropertyTypes" as const,
-    primaryKeyApiName: "id" as const,
-    primaryKeyType: "integer",
-    links: {},
-    icon: {
-      type: "blueprint",
-      name: "icon",
-      color: "blue",
-    },
-    properties: {
-      id: {
-        multiplicity: false,
-        type: "integer",
-      },
-      string: {
-        multiplicity: false,
-        type: "string",
-      },
-      boolean: {
-        multiplicity: false,
-        type: "boolean",
-      },
-      date: {
-        multiplicity: false,
-        type: "datetime",
-      },
-      dateTime: {
-        multiplicity: false,
-        type: "timestamp",
-      },
-      decimal: {
-        multiplicity: false,
-        type: "decimal",
-      },
-      integer: {
-        multiplicity: false,
-        type: "integer",
-      },
-      long: {
-        multiplicity: false,
-        type: "long",
-      },
-      short: {
-        multiplicity: false,
-        type: "short",
-      },
-      float: {
-        multiplicity: false,
-        type: "float",
-      },
-      double: {
-        multiplicity: false,
-        type: "double",
-      },
-      byte: {
-        multiplicity: false,
-        type: "byte",
-      },
-      attachment: {
-        multiplicity: false,
-        type: "attachment",
-      },
-      geoPoint: {
-        multiplicity: false,
-        type: "geopoint",
-      },
-      geoShape: {
-        multiplicity: false,
-        type: "geoshape",
-      },
-      stringArray: {
-        multiplicity: true,
-        type: "string",
-      },
-      booleanArray: {
-        multiplicity: true,
-        type: "boolean",
-      },
-      dateArray: {
-        multiplicity: true,
-        type: "datetime",
-      },
-      dateTimeArray: {
-        multiplicity: true,
-        type: "timestamp",
-      },
-      decimalArray: {
-        multiplicity: true,
-        type: "decimal",
-      },
-      integerArray: {
-        multiplicity: true,
-        type: "integer",
-      },
-      longArray: {
-        multiplicity: true,
-        type: "long",
-      },
-      shortArray: {
-        multiplicity: true,
-        type: "short",
-      },
-      floatArray: {
-        multiplicity: true,
-        type: "float",
-      },
-      doubleArray: {
-        multiplicity: true,
-        type: "double",
-      },
-      byteArray: {
-        multiplicity: true,
-        type: "byte",
-      },
-      attachmentArray: {
-        multiplicity: true,
-        type: "attachment",
-      },
-      geoPointArray: {
-        multiplicity: true,
-        type: "geopoint",
-      },
-      geoShapeArray: {
-        multiplicity: true,
-        type: "geoshape",
-      },
-      mediaReference: {
-        multiplicity: false,
-        type: "mediaReference",
-      },
-    },
-    titleProperty: "id" as const,
-    displayName: "ObjectTypeWithAllPropertyTypes",
-    pluralDisplayName: "ObjectTypeWithAllPropertyTypes",
-    status: "ACTIVE",
-    interfaceMap: {},
-    inverseInterfaceMap: {},
-    rid: "rid.object-type.123",
-  } satisfies ObjectMetadata;
